+++ packages/fields/tests/Implementation.test.js
@@ -7,6 +7,7 @@ const args = {
     newFieldAdapter: jest.fn(),
   },
   defaultAccess: true,
+  schemaNames: ['public'],
 };
 
 describe('new Implementation()', () => {
@@ -21,6 +22,7 @@ describe('new Implementation()', () => {
           newFieldAdapter: jest.fn(),
         },
         defaultAccess: true,
+        schemaNames: ['public'],
       }
     );
     expect(impl).not.toBeNull();
@@ -46,14 +48,14 @@ test('addToMongooseSchema()', () => {
 
 test('getGqlAuxTypes()', () => {
   const impl = new Field('path', {}, args);
-
-  expect(impl.getGqlAuxTypes()).toEqual([]);
+  const schemaName = 'public';
+  expect(impl.getGqlAuxTypes({ schemaName })).toEqual([]);
 });
 
 test('gqlAuxFieldResolvers', () => {
   const impl = new Field('path', {}, args);
-
-  expect(impl.gqlAuxFieldResolvers).toEqual({});
+  const schemaName = 'public';
+  expect(impl.gqlAuxFieldResolvers({ schemaName })).toEqual({});
 });
 
 test('getGqlAuxQueries()', () => {
@@ -65,7 +67,7 @@ test('getGqlAuxQueries()', () => {
 test('gqlAuxQueryResolvers', () => {
   const impl = new Field('path', {}, args);
 
-  expect(impl.gqlAuxQueryResolvers).toEqual({});
+  expect(impl.gqlAuxQueryResolvers()).toEqual({});
 });
 
 test('getGqlAuxMutations()', () => {
@@ -77,7 +79,7 @@ test('getGqlAuxMutations()', () => {
 test('gqlAuxMutationResolvers', () => {
   const impl = new Field('path', {}, args);
 
-  expect(impl.gqlAuxMutationResolvers).toEqual({});
+  expect(impl.gqlAuxMutationResolvers()).toEqual({});
 });
 
 test('afterChange()', async () => {
@@ -97,8 +99,8 @@ test('resolveInput()', async () => {
 
 test('gqlQueryInputFields', () => {
   const impl = new Field('path', {}, args);
-
-  expect(impl.gqlQueryInputFields).toEqual([]);
+  const schemaName = 'public';
+  expect(impl.gqlQueryInputFields({ schemaName })).toEqual([]);
 });
 
 test('gqlUpdateInputFields', () => {
@@ -109,15 +111,16 @@ test('gqlUpdateInputFields', () => {
 
 test('gqlOutputFieldResolvers', () => {
   const impl = new Field('path', {}, args);
-
-  expect(impl.gqlOutputFieldResolvers).toEqual({});
+  const schemaName = 'public';
+  expect(impl.gqlOutputFieldResolvers({ schemaName })).toEqual({});
 });
 
 describe('getAdminMeta()', () => {
   test('meta is as expect', () => {
     const impl = new Field('path', { label: 'config label', defaultValue: 'default' }, args);
+    const schemaName = 'public';
 
-    const value = impl.getAdminMeta();
+    const value = impl.getAdminMeta({ schemaName });
     expect(value).toEqual({
       access: {
         create: true,
@@ -135,8 +138,9 @@ describe('getAdminMeta()', () => {
 
   test('when defaultValue is a function, forced to `undefined`', () => {
     const impl = new Field('path', { label: 'config label', defaultValue: () => 'default' }, args);
+    const schemaName = 'public';
 
-    const value = impl.getAdminMeta();
+    const value = impl.getAdminMeta({ schemaName });
     expect(value).toEqual({
       access: {
         create: true,