+++ packages/auth-passport/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @keystone-alpha/auth-passport
 
+## 4.1.2
+
+- Updated dependencies [9ade2b2d](https://github.com/keystonejs/keystone-5/commit/9ade2b2d):
+  - @keystone-alpha/fields@12.0.0
+
+## 4.1.1
+
+- Updated dependencies [89c0d7e9](https://github.com/keystonejs/keystone-5/commit/89c0d7e9):
+- Updated dependencies [a8e9378d](https://github.com/keystonejs/keystone-5/commit/a8e9378d):
+  - @keystone-alpha/fields@11.0.0
+
 ## 4.1.0
 
 ### Minor Changes