+++ packages/fields-auto-increment/CHANGELOG.md
@@ -1,5 +1,28 @@
 # @keystone-alpha/fields-auto-increment
 
+## 2.0.1
+
+- Updated dependencies [9ade2b2d](https://github.com/keystonejs/keystone-5/commit/9ade2b2d):
+  - @keystone-alpha/adapter-knex@4.0.9
+  - @keystone-alpha/fields@12.0.0
+
+## 2.0.0
+
+### Major Changes
+
+- [89c0d7e9](https://github.com/keystonejs/keystone-5/commit/89c0d7e9): The `.access` property of Fields is now keyed by `schemaName`. As such, a number of getters have been replaced with methods which take `{ schemaName }`.
+
+  - `get gqlOutputFields()` -> `gqlOutputFields({ schemaName })`
+  - `get gqlOutputFieldResolvers()` -> `gqlOutputFieldResolvers({ schemaName })`
+  - `get gqlAuxFieldResolvers() -> gqlAuxFieldResolvers({ schemaName })`
+  - `get gqlAuxQueryResolvers()` -> `gqlAuxQueryResolvers({ schemaName })`
+  - `get gqlAuxMutationResolvers()` -> `gqlAuxMutationResolvers({ schemaName })`
+  - `get gqlQueryInputFields()` -> `gqlQueryInputFields({ schemaName })`
+
+* Updated dependencies [a8e9378d](https://github.com/keystonejs/keystone-5/commit/a8e9378d):
+  - @keystone-alpha/adapter-knex@4.0.8
+  - @keystone-alpha/fields@11.0.0
+
 ## 1.0.1
 
 - Updated dependencies [144e6e86](https://github.com/keystonejs/keystone-5/commit/144e6e86):